02_p_n_n.py & 02_bmi angepasst

+++ Aufgabe 3.2.4/Einsteiger/02_p_n_n.py
@@ -2,7 +2,7 @@
 
 print("Positiv, Negativ oder Null?")
 
-number = int(input("Gebe ein Zahl ein: "))
+number = float(input("Gebe ein Zahl ein: "))
 
 if number > 0:
     print("Diese Zahl ist positiv")
